+++ src/main/java/pl/allegro/finance/tradukisto/MoneyConverters.java
@@ -1,30 +1,36 @@
 package pl.allegro.finance.tradukisto;
 
-import static com.google.common.base.Verify.verifyNotNull;
-import static pl.allegro.finance.tradukisto.internal.Container.brazilianPortugueseContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.czechContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.englishContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.germanContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.polishContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.russianContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.serbianCyrillicContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.slovakContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.latvianContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.kazakhContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.ukrainianContainer;
-import static pl.allegro.finance.tradukisto.internal.Container.serbianContainer;
+import pl.allegro.finance.tradukisto.internal.BigDecimalToStringConverter;
 
 import java.math.BigDecimal;
 
-import pl.allegro.finance.tradukisto.internal.BigDecimalToStringConverter;
+import static com.google.common.base.Verify.verifyNotNull;
+import static pl.allegro.finance.tradukisto.internal.Container.americanEnglishContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.brazilianPortugueseContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.czechContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.englishContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.frenchContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.germanContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.italianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.kazakhContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.latvianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.polishContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.russianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.serbianCyrillicContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.serbianContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.slovakContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.turkishContainer;
+import static pl.allegro.finance.tradukisto.internal.Container.ukrainianContainer;
 
 public enum MoneyConverters {
 
     BRAZILIAN_PORTUGUESE_BANKING_MONEY_VALUE(brazilianPortugueseContainer().getBankingMoneyConverter()),
     ENGLISH_BANKING_MONEY_VALUE(englishContainer().getBankingMoneyConverter()),
+    AMERICAN_ENGLISH_BANKING_MONEY_VALUE(americanEnglishContainer().getBankingMoneyConverter()),
     GERMAN_BANKING_MONEY_VALUE(germanContainer().getBankingMoneyConverter()),
     RUSSIAN_BANKING_MONEY_VALUE(russianContainer().getBankingMoneyConverter()),
     POLISH_BANKING_MONEY_VALUE(polishContainer().getBankingMoneyConverter()),
+    ITALIAN_BANKING_MONEY_VALUE(italianContainer().getBankingMoneyConverter()),
     CZECH_BANKING_MONEY_VALUE(czechContainer().getBankingMoneyConverter()),
     SLOVAK_BANKING_MONEY_VALUE(slovakContainer().getBankingMoneyConverter()),
     LATVIAN_BANKING_MONEY_VALUE(latvianContainer().getBankingMoneyConverter()),
@@ -32,6 +38,8 @@ public enum MoneyConverters {
     UKRAINIAN_BANKING_MONEY_VALUE(ukrainianContainer().getBankingMoneyConverter()),
     SERBIAN_BANKING_MONEY_VALUE(serbianContainer().getBankingMoneyConverter()),
     SERBIAN_CYRILLIC_BANKING_MONEY_VALUE(serbianCyrillicContainer().getBankingMoneyConverter());
+    TURKISH_BANKING_MONEY_VALUE(turkishContainer().getBankingMoneyConverter()),
+    FRENCH_BANKING_MONEY_VALUE(frenchContainer().getBankingMoneyConverter());
 
     private final BigDecimalToStringConverter converter;