+++ src/main/java/pl/allegro/finance/tradukisto/internal/Container.java
@@ -6,12 +6,16 @@ import pl.allegro.finance.tradukisto.internal.converters.IntegerToWordsConverter
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValues;
 import pl.allegro.finance.tradukisto.internal.languages.czech.CzechValuesForSmallNumbers;
+import pl.allegro.finance.tradukisto.internal.languages.english.AmericanEnglishValues;
 import pl.allegro.finance.tradukisto.internal.languages.english.EnglishValues;
 import pl.allegro.finance.tradukisto.internal.languages.french.FrenchIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.french.FrenchValues;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanIntegerToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanThousandToWordsConverter;
 import pl.allegro.finance.tradukisto.internal.languages.german.GermanValues;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianThousandToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.italian.ItalianValues;
 import pl.allegro.finance.tradukisto.internal.languages.kazakh.KazakhValues;
 import pl.allegro.finance.tradukisto.internal.languages.latvian.LatvianValues;
 import pl.allegro.finance.tradukisto.internal.languages.polish.PolishValues;
@@ -24,6 +28,10 @@ import pl.allegro.finance.tradukisto.internal.languages.serbian.SerbianCyrillicV
 import pl.allegro.finance.tradukisto.internal.languages.serbian.SerbianValues;
 import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValues;
 import pl.allegro.finance.tradukisto.internal.languages.slovak.SlovakValuesForSmallNumbers;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishBigDecimalToBankingMoneyConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishIntegerToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishSmallNumbersToWordsConverter;
+import pl.allegro.finance.tradukisto.internal.languages.turkish.TurkishValues;
 import pl.allegro.finance.tradukisto.internal.languages.ukrainian.UkrainianValues;
 
 public final class Container {
@@ -81,6 +89,10 @@ public final class Container {
         return new Container(new EnglishValues());
     }
 
+    public static Container americanEnglishContainer() {
+        return new Container(new AmericanEnglishValues());
+    }
+
     public static Container frenchContainer() {
 
         FrenchValues values = new FrenchValues();
@@ -115,6 +127,22 @@ public final class Container {
         return new Container(converter, bigDecimalBankingMoneyValueConverter);
     }
 
+    public static Container italianContainer() {
+        ItalianValues values = new ItalianValues();
+
+        ItalianThousandToWordsConverter italianThousandToWordsConverter = new ItalianThousandToWordsConverter(
+                values.baseNumbers());
+
+        IntegerToStringConverter converter = new ItalianIntegerToWordsConverter(
+                new IntegerToWordsConverter(italianThousandToWordsConverter, values.pluralForms()), values.exceptions(),
+                italianThousandToWordsConverter);
+
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(
+                converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
     public static Container latvianContainer() {
         return new Container(new LatvianValues());
     }
@@ -131,6 +159,19 @@ public final class Container {
 
         BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter = new BigDecimalToBankingMoneyConverter(
                 converter, values.currency());
+
+        return new Container(converter, bigDecimalBankingMoneyValueConverter);
+    }
+
+    public static Container turkishContainer() {
+        TurkishValues values = new TurkishValues();
+
+        TurkishSmallNumbersToWordsConverter smallNumbersConverter = new TurkishSmallNumbersToWordsConverter(values);
+        IntegerToWordsConverter bigNumbersConverter = new IntegerToWordsConverter(smallNumbersConverter, values.pluralForms());
+        IntegerToStringConverter converter =
+                new TurkishIntegerToWordsConverter(bigNumbersConverter, smallNumbersConverter);
+        BigDecimalToStringConverter bigDecimalBankingMoneyValueConverter =
+                new TurkishBigDecimalToBankingMoneyConverter(converter, values);
 
         return new Container(converter, bigDecimalBankingMoneyValueConverter);
     }