+++ src/main/java/pl/allegro/finance/tradukisto/internal/languages/english/AmericanEnglishValues.java
@@ -1,0 +1,8 @@
+package pl.allegro.finance.tradukisto.internal.languages.english;
+
+public class AmericanEnglishValues extends EnglishValues {
+    @Override
+    public String currency() {
+        return "$";
+    }
+}