+++ src/test/groovy/pl/allegro/finance/tradukisto/MoneyConvertersTest.groovy
@@ -3,15 +3,20 @@ package pl.allegro.finance.tradukisto
 import com.google.common.base.VerifyException
 import spock.lang.Specification
 
+import static pl.allegro.finance.tradukisto.MoneyConverters.AMERICAN_ENGLISH_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.BRAZILIAN_PORTUGUESE_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.CZECH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.ENGLISH_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.FRENCH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.GERMAN_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.ITALIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.KAZAKH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.LATVIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.POLISH_BANKING_MONEY_VALUE
-import static pl.allegro.finance.tradukisto.MoneyConverters.BRAZILIAN_PORTUGUESE_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.RUSSIAN_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.SERBIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.SLOVAK_BANKING_MONEY_VALUE
+import static pl.allegro.finance.tradukisto.MoneyConverters.TURKISH_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.UKRAINIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.SERBIAN_BANKING_MONEY_VALUE
 import static pl.allegro.finance.tradukisto.MoneyConverters.SERBIAN_CYRILLIC_BANKING_MONEY_VALUE
@@ -26,6 +31,11 @@ class MoneyConvertersTest extends Specification {
     def "should convert money in German"() {
         expect:
         GERMAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "eintausendzweihundertvierunddreißig € 56/100"
+    }
+
+    def "should convert money in Italian"() {
+        expect:
+        ITALIAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "milleduecentotrentaquattro € 56/100"
     }
 
     def "should convert money in Russian"() {
@@ -53,9 +63,14 @@ class MoneyConvertersTest extends Specification {
         LATVIAN_BANKING_MONEY_VALUE.asWords(1_234.56) == "viens tūkstotis divi simti trīsdesmit četri EUR 56/100"
     }
 
-    def "should convert money in English"() {
+    def "should convert money in British English"() {
         expect:
         ENGLISH_BANKING_MONEY_VALUE.asWords(1_234.56) == "one thousand two hundred thirty-four £ 56/100"
+    }
+
+    def "should convert money in American English"() {
+        expect:
+        AMERICAN_ENGLISH_BANKING_MONEY_VALUE.asWords(1_234.56) == "one thousand two hundred thirty-four \$ 56/100"
     }
 
     def "should convert money in Kazakh"() {
@@ -78,11 +93,21 @@ class MoneyConvertersTest extends Specification {
         SERBIAN_CYRILLIC_BANKING_MONEY_VALUE.asWords(1_234.56) == "једна хиљада двеста тридесет четири РСД 56/100"
     }
 
+    def "should convert numbers in French"() {
+        expect:
+        FRENCH_BANKING_MONEY_VALUE.asWords(1_234.56) == "mille deux cent trente-quatre € 56/100"
+    }
+
     def "should throw exception when null given"() {
         when:
         POLISH_BANKING_MONEY_VALUE.asWords(null)
 
         then:
         thrown(VerifyException)
+    }
+
+    def "should convert money in Turkish"() {
+        expect:
+        TURKISH_BANKING_MONEY_VALUE.asWords(1_23.4) == "YüzYirmiÜçTL,KırkKr."
     }
 }