+++ src/test/groovy/pl/allegro/finance/tradukisto/ValueConvertersTest.groovy
@@ -3,15 +3,7 @@ package pl.allegro.finance.tradukisto
 import com.google.common.base.VerifyException
 import spock.lang.Specification
 
-import static pl.allegro.finance.tradukisto.ValueConverters.CZECH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.ENGLISH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.GERMAN_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.BRAZILIAN_PORTUGUESE_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.KAZAKH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.POLISH_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.RUSSIAN_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.LATVIAN_INTEGER
-import static pl.allegro.finance.tradukisto.ValueConverters.FRENCH_INTEGER
+import static pl.allegro.finance.tradukisto.ValueConverters.*
 
 class ValueConvertersTest extends Specification {
 
@@ -23,6 +15,11 @@ class ValueConvertersTest extends Specification {
     def "should convert numbers in German"() {
         expect:
         GERMAN_INTEGER.asWords(1_234) == "eintausendzweihundertvierunddreißig"
+    }
+
+    def "should convert numbers in Italian"() {
+        expect:
+        ITALIAN_INTEGER.asWords(1_234) == "milleduecentotrentaquattro"
     }
 
     def "should convert numbers in Russian"() {
@@ -53,6 +50,11 @@ class ValueConvertersTest extends Specification {
     def "should convert numbers in Kazakh"() {
         expect:
         KAZAKH_INTEGER.asWords(1_234) == "бір мың екі жүз отыз төрт"
+    }
+
+    def "should convert numbers in Turkish"() {
+        expect:
+        TURKISH_INTEGER.asWords(1_234) == "Bin İki Yüz Otuz Dört"
     }
 
     def "should convert numbers in French"() {